#SB-250: Create a service that allows configuring the logging programatically
- Added in Logback file validation and log validation
- Converted LogXMLConfiguration class static to non-static
