+++ logging-configuration-rest/src/test/java/org/carlspring/logging/rest/LogGenerator.java
@@ -1,0 +1,21 @@
+package org.carlspring.logging.rest;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class LogGenerator
+{
+
+    private Logger logger = LoggerFactory.getLogger(LogGenerator.class);
+
+    public void infoLog()
+    {
+        logger.info("info log");
+    }
+
+    public void debugLog()
+    {
+        logger.debug("debug log");
+    }
+
+}
